refactor: move createPersonaByMnemonic
