+++ packages/mask/background/services/identity/index.ts
@@ -1,4 +1,4 @@
-export { createPersonaByPrivateKey, mobile_restoreFromMnemonicWords } from './persona/create'
+export { createPersonaByPrivateKey, mobile_restoreFromMnemonicWords, createPersonaByMnemonic } from './persona/create'
 export { mobile_queryPersonaRecordsFromIndexedDB } from './persona/query'
 export {
     deletePersona,
